move aws cycles to test package
